Revert "Merge pull request #1154 from deathbeam/team-member-names"
This reverts commit f0293087ce0a2e2dc515cfb57ef926c185d25cb5, reversing
changes made to 309c1fde25d32086cfd242f988c8f45df903d66c.
