--- runelite-api/src/main/java/net/runelite/api/Text.java
@@ -1,65 +1,0 @@
-/*
- * Copyright (c) 2018, Joshua Filby <joshua@filby.me>
- * All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- *
- * 1. Redistributions of source code must retain the above copyright notice, this
- *    list of conditions and the following disclaimer.
- * 2. Redistributions in binary form must reproduce the above copyright notice,
- *    this list of conditions and the following disclaimer in the documentation
- *    and/or other materials provided with the distribution.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
- * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
- * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
- * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
- * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
- * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
- * ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
- * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-package net.runelite.api;
-
-/**
- * A set of utilities to use when dealing with text.
- */
-public class Text
-{
-
-	/**
-	 * Removes all tags from the given `str`.
-	 *
-	 * @param str The string to remove tags from.
-	 * @return The given `str` with all tags removed from it.
-	 */
-	public static String removeTags(String str)
-	{
-		StringBuilder builder = new StringBuilder(str.length());
-		boolean inTag = false;
-
-		for (int i = 0; i < str.length(); i++)
-		{
-			char currentChar = str.charAt(i);
-
-			if (currentChar == '<')
-			{
-				inTag = true;
-			}
-			else if (currentChar == '>')
-			{
-				inTag = false;
-			}
-			else if (!inTag)
-			{
-				builder.append(currentChar);
-			}
-		}
-
-		return builder.toString();
-	}
-
-}
