+++ runelite-client/src/main/java/net/runelite/client/plugins/hiscore/HiscorePlugin.java
@@ -32,7 +32,6 @@ import java.util.concurrent.ScheduledExecutorService;
 import javax.imageio.ImageIO;
 import javax.inject.Inject;
 import javax.swing.SwingUtilities;
-import net.runelite.api.Text;
 import net.runelite.api.events.ConfigChanged;
 import net.runelite.api.events.PlayerMenuOptionClicked;
 import net.runelite.client.config.ConfigManager;
@@ -139,7 +138,7 @@ public class HiscorePlugin extends Plugin
 					throw new RuntimeException(e);
 				}
 
-				hiscorePanel.lookup(Text.removeTags(event.getMenuTarget()));
+				hiscorePanel.lookup(event.getMenuTarget());
 			});
 		}
 	}
