+++ runelite-client/src/main/java/net/runelite/client/util/Text.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2018, Tomas Slusny <slusnucky@gmail.com>
+ * Copyright (c) 2018, Joshua Filby <joshua@filby.me>
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -22,12 +22,44 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package net.runelite.api.events;
+package net.runelite.client.util;
 
-import lombok.Value;
-
-@Value
-public class ClanChanged
+/**
+ * A set of utilities to use when dealing with text.
+ */
+public class Text
 {
-	private boolean joined;
+
+	/**
+	 * Removes all tags from the given `str`.
+	 *
+	 * @param str The string to remove tags from.
+	 * @return The given `str` with all tags removed from it.
+	 */
+	public static String removeTags(String str)
+	{
+		StringBuilder builder = new StringBuilder(str.length());
+		boolean inTag = false;
+
+		for (int i = 0; i < str.length(); i++)
+		{
+			char currentChar = str.charAt(i);
+
+			if (currentChar == '<')
+			{
+				inTag = true;
+			}
+			else if (currentChar == '>')
+			{
+				inTag = false;
+			}
+			else if (!inTag)
+			{
+				builder.append(currentChar);
+			}
+		}
+
+		return builder.toString();
+	}
+
 }
