+++ runelite-client/src/main/resources/net/runelite/client/plugins/clanchat/Captain_clan_rank.png
