+++ runelite-client/src/main/resources/net/runelite/client/plugins/clanchat/Corporal_clan_rank.png
