+++ runelite-client/src/main/resources/net/runelite/client/plugins/clanchat/Friend_clan_rank.png
