+++ runelite-client/src/main/resources/net/runelite/client/plugins/clanchat/General_clan_rank.png
