+++ runelite-client/src/main/resources/net/runelite/client/plugins/clanchat/Lieutenant_clan_rank.png
