+++ runelite-client/src/main/resources/net/runelite/client/plugins/clanchat/Owner_clan_rank.png
