+++ runelite-client/src/main/resources/net/runelite/client/plugins/clanchat/Recruit_clan_rank.png
