+++ runelite-client/src/main/resources/net/runelite/client/plugins/clanchat/Sergeant_clan_rank.png
