+++ runelite-client/src/test/java/net/runelite/client/util/TextTest.java
@@ -24,7 +24,6 @@
  */
 package net.runelite.client.util;
 
-import net.runelite.api.Text;
 import static org.junit.Assert.assertEquals;
 import org.junit.Test;
