+++ runelite-mixins/src/main/java/net/runelite/mixins/RSClientMixin.java
@@ -56,7 +56,6 @@ import net.runelite.api.Varbits;
 import net.runelite.api.WidgetNode;
 import net.runelite.api.coords.LocalPoint;
 import net.runelite.api.events.BoostedLevelChanged;
-import net.runelite.api.events.ClanChanged;
 import net.runelite.api.events.DraggingWidgetChanged;
 import net.runelite.api.events.ExperienceChanged;
 import net.runelite.api.events.GameStateChanged;
@@ -653,12 +652,5 @@ public abstract class RSClientMixin implements RSClient
 		ResizeableChanged resizeableChanged = new ResizeableChanged();
 		resizeableChanged.setResized(client.isResized());
 		eventBus.post(resizeableChanged);
-	}
-
-	@FieldHook("clanMemberManager")
-	@Inject
-	public static void clanMemberManagerChanged(int idx)
-	{
-		eventBus.post(new ClanChanged(client.getClanMemberManager() != null));
 	}
 }
