+++ runescape-api/src/main/java/net/runelite/rs/api/RSClient.java
@@ -549,12 +549,4 @@ public interface RSClient extends RSGameEngine, Client
 
 	@Import("drawObject")
 	void drawObject(int z, int x, int y, int randomColor1, int randomColor2);
-
-	@Import("playerNameMask")
-	@Override
-	void setPlayerNameMask(int mask);
-
-	@Import("playerNameMask")
-	@Override
-	int getPlayerNameMask();
 }
