+++ runescape-api/src/main/java/net/runelite/rs/api/RSName.java
@@ -30,10 +30,4 @@ public interface RSName
 {
 	@Import("name")
 	String getName();
-
-	@Import("cleanName")
-	String getCleanName();
-
-	@Import("name")
-	void setName(String name);
 }
