+++ runescape-client/src/main/java/Client.java
@@ -951,8 +951,7 @@ public final class Client extends GameEngine implements class302 {
    @ObfuscatedGetter(
       intValue = -1162998275
    )
-   @Export("playerNameMask")
-   static int playerNameMask;
+   static int field975;
    @ObfuscatedName("js")
    @ObfuscatedGetter(
       intValue = -764738135
@@ -1334,7 +1333,7 @@ public final class Client extends GameEngine implements class302 {
       localInteractingIndex = -1;
       field973 = 0;
       field974 = true;
-      playerNameMask = 0;
+      field975 = 0;
       field976 = 0;
       field995 = new int[1000];
       playerMenuTypes = new int[]{44, 45, 46, 47, 48, 49, 50, 51};
